AI-Driven Detection of Deepfake Content for Enhancing Digital Trust/requirements.txt
"""flask==3.0.0
flask-cors==4.0.0

torch==2.1.0
torchvision==0.16.0

tensorflow==2.15.0
keras==2.15.0

opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0

numpy==1.26.4"""